Fix PCM32bit output crackling. This commit fixes ticket #988
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@2789 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/transform/MpaDecFilter/MpaDecFilter.cpp
@@ -1882,7 +1882,7 @@ HRESULT CMpaDecFilter::Deliver(CAtlArray<float>& pBuff, DWORD nSamplesPerSec, WO
 			}
 			break;
 			case SF_PCM32:
-				*(int*)pDataOut = (int)round(f * INT_MAX);
+				*(int*)pDataOut = (int)round((double)f * INT_MAX);
 				pDataOut += sizeof(int);
 				break;
 			case SF_FLOAT32:
